pyproject.toml
[project]
name = "swarmhq"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-adk>=1.28.1",
]
